convert point group to a backbone model.
everything seems to be working, but this might have introduced bugs. the rationale is that we can now make the list of point groups a backbone collection, with a corresponding view, and do view updating after group removal / reordering via the event system.

updating points in the group needs to go via the methods provided, rather than .set('point_ids', [..]) because the model object stores the point ids in a map so that membership can be tested efficiently, and sets the point_ids attribute to a sorted list of keys of that map (updating as changes are made). however, the same isn't done in reverse, so if the point_ids attribute is set externally, then it will be out of sync with the map.