Support cases (such as TCGA) where the sample annotation IDs do not match the dataset IDs.
Where there is a complete mismatch (no IDs in common), search for a mapping from sample IDs to dat set IDs. The two alternatives supported are: 

* sample IDs are prefixes of the dataset IDs (prefix_map)
* dataset IDs are prefixes of the sample IDs (rev_prefix_map)

if a mapping is found that maps sample IDs unambiguously to dataset IDs (allowing some sample IDs not to be mapped, but requiring that all dataset IDs must map), then a copy is made of the sample annotation, and the sample IDs are remapped to dataset IDs.

it might be worth considering whether to always make a copy, and reduce the set of annotations down to the intersection with the dataset. this has an effect on over-representation testing.