Know bug:  the transformation is not applied on the x-axis.
Fixed current selection that was not dispalyed after resize(issue20)
Fixed tickValue error causing problem for the transition
Display the transformation used in the textpanel
Modify the colours for the group avoiding to start at 0 when a box is already present.
Added an alert message (needs to be nicer) when user select a color and save without pressing outside the spectrum selector.
Fixed some page behaviors related to user interacting with the page